detect: move generic int detect function
from http2 to a generic file
so that it can be reused by dcerpc and others
