+++ rust/src/http2/detect.rs
@@ -20,6 +20,7 @@ use super::http2::{
 };
 use super::parser;
 use crate::core::Direction;
+use crate::detect::{detect_parse_u64, DetectU64Data, DetectUintMode};
 use std::ffi::CStr;
 use std::str::FromStr;
 
@@ -255,22 +256,22 @@ fn http2_detect_settings_match(
                     return 1;
                 }
                 Some(x) => match x.mode {
-                    parser::DetectUintMode::DetectUintModeEqual => {
+                    DetectUintMode::DetectUintModeEqual => {
                         if set[i].value == x.value {
                             return 1;
                         }
                     }
-                    parser::DetectUintMode::DetectUintModeLt => {
+                    DetectUintMode::DetectUintModeLt => {
                         if set[i].value <= x.value {
                             return 1;
                         }
                     }
-                    parser::DetectUintMode::DetectUintModeGt => {
+                    DetectUintMode::DetectUintModeGt => {
                         if set[i].value >= x.value {
                             return 1;
                         }
                     }
-                    parser::DetectUintMode::DetectUintModeRange => {
+                    DetectUintMode::DetectUintModeRange => {
                         if set[i].value <= x.value && set[i].value >= x.valrange {
                             return 1;
                         }
@@ -326,7 +327,7 @@ pub unsafe extern "C" fn rs_detect_u64_parse(
 ) -> *mut std::os::raw::c_void {
     let ft_name: &CStr = CStr::from_ptr(str); //unsafe
     if let Ok(s) = ft_name.to_str() {
-        if let Ok((_, ctx)) = parser::detect_parse_u64(s) {
+        if let Ok((_, ctx)) = detect_parse_u64(s) {
             let boxed = Box::new(ctx);
             return Box::into_raw(boxed) as *mut _;
         }
@@ -337,36 +338,36 @@ pub unsafe extern "C" fn rs_detect_u64_parse(
 #[no_mangle]
 pub unsafe extern "C" fn rs_detect_u64_free(ctx: *mut std::os::raw::c_void) {
     // Just unbox...
-    std::mem::drop(Box::from_raw(ctx as *mut parser::DetectU64Data));
+    std::mem::drop(Box::from_raw(ctx as *mut DetectU64Data));
 }
 
 fn http2_detect_sizeupdate_match(
-    blocks: &[parser::HTTP2FrameHeaderBlock], ctx: &parser::DetectU64Data,
+    blocks: &[parser::HTTP2FrameHeaderBlock], ctx: &DetectU64Data,
 ) -> std::os::raw::c_int {
     for block in blocks.iter() {
         match ctx.mode {
-            parser::DetectUintMode::DetectUintModeEqual => {
+            DetectUintMode::DetectUintModeEqual => {
                 if block.sizeupdate == ctx.value
                     && block.error == parser::HTTP2HeaderDecodeStatus::HTTP2HeaderDecodeSizeUpdate
                 {
                     return 1;
                 }
             }
-            parser::DetectUintMode::DetectUintModeLt => {
+            DetectUintMode::DetectUintModeLt => {
                 if block.sizeupdate <= ctx.value
                     && block.error == parser::HTTP2HeaderDecodeStatus::HTTP2HeaderDecodeSizeUpdate
                 {
                     return 1;
                 }
             }
-            parser::DetectUintMode::DetectUintModeGt => {
+            DetectUintMode::DetectUintModeGt => {
                 if block.sizeupdate >= ctx.value
                     && block.error == parser::HTTP2HeaderDecodeStatus::HTTP2HeaderDecodeSizeUpdate
                 {
                     return 1;
                 }
             }
-            parser::DetectUintMode::DetectUintModeRange => {
+            DetectUintMode::DetectUintModeRange => {
                 if block.sizeupdate <= ctx.value
                     && block.sizeupdate >= ctx.valrange
                     && block.error == parser::HTTP2HeaderDecodeStatus::HTTP2HeaderDecodeSizeUpdate
@@ -396,7 +397,7 @@ fn http2_header_blocks(frame: &HTTP2Frame) -> Option<&[parser::HTTP2FrameHeaderB
 }
 
 fn http2_detect_sizeupdatectx_match(
-    ctx: &mut parser::DetectU64Data, tx: &mut HTTP2Transaction, direction: Direction,
+    ctx: &mut DetectU64Data, tx: &mut HTTP2Transaction, direction: Direction,
 ) -> std::os::raw::c_int {
     if direction == Direction::ToServer {
         for i in 0..tx.frames_ts.len() {
@@ -422,7 +423,7 @@ fn http2_detect_sizeupdatectx_match(
 pub unsafe extern "C" fn rs_http2_detect_sizeupdatectx_match(
     ctx: *const std::os::raw::c_void, tx: *mut std::os::raw::c_void, direction: u8,
 ) -> std::os::raw::c_int {
-    let ctx = cast_pointer!(ctx, parser::DetectU64Data);
+    let ctx = cast_pointer!(ctx, DetectU64Data);
     let tx = cast_pointer!(tx, HTTP2Transaction);
     return http2_detect_sizeupdatectx_match(ctx, tx, direction.into());
 }
